fix: resolve Gradle plugin compatibility and Dart type errors
- Regenerate platform files after pub get (flutter create --overwrite)
- Pin package_info_plus to ^8.0.0 (9.0.1 has Gradle plugin issues)
- Fix num-to-double type errors in player_controls.dart
- Remove broken analysis_options.yaml include
- Remove auto-generated test file

+++ lib/widgets/player_controls.dart
@@ -59,7 +59,7 @@ class PlayerControls extends StatelessWidget {
     }
 
     final posMs = position.inMilliseconds.toDouble();
-    final durMs = duration.inMilliseconds.toDouble().clamp(1, double.infinity);
+    final durMs = duration.inMilliseconds.toDouble().clamp(1.0, double.infinity).toDouble();
 
     return Container(
       decoration: const BoxDecoration(
@@ -124,8 +124,8 @@ class PlayerControls extends StatelessWidget {
                     overlayShape: const RoundSliderOverlayShape(overlayRadius: 14),
                   ),
                   child: Slider(
-                    value: posMs.clamp(0, durMs),
-                    min: 0,
+                    value: posMs.clamp(0.0, durMs).toDouble(),
+                    min: 0.0,
                     max: durMs,
                     onChanged: onSeekChanged,
                   ),
